Avoid a type-sizing warning between gulong and gsize
This seems to happen on Windows, according to Wireshark's CI.

+++ communityid.c
@@ -342,7 +342,7 @@ static gboolean communityid_calc(communityid_cfg_t *cfg, guint8 proto,
         gsize len = strlen(CID_VERSION_PREFIX) + strlen(str) + 1;
 
         *result = (gchar*) g_malloc(len);
-        g_snprintf(*result, len, "%s%s", CID_VERSION_PREFIX, str);
+        g_snprintf(*result, (gulong) len, "%s%s", CID_VERSION_PREFIX, str);
         g_free(str);
     } else {
         /* Convert binary SHA-1 to ASCII representation.
